Issue #033: fix: remove EP registration from unit test main (#144)
## Fix: EP Duplicate Registration Causing V2 API Test Failure

### Problem
V2 API E2E test fails when attempting to register
MorphiZenExecutionProvider because the EP was already registered
globally in main() function. ONNX Runtime does not allow duplicate
registration of EP library with same name.

### Solution
Remove EP registration from unit test main and use
setup_global_morphizen_ort_api() for unit tests instead. This allows E2E
tests to register EP independently without duplicate registration error.

### Changes
- Remove Ort::Env and RegisterExecutionProviderLibrary from
unit-test/morphizen_unit_test_main.cpp
- Unit tests now use setup_global_morphizen_ort_api() for initialization
- E2E tests can register EP independently
- Apply linting fixes
- Update backlog documentation

Fixes Issue #033

+++ morphizen-core/src/main.cpp
@@ -28,7 +28,7 @@ static struct {
 };
 
 static void* lookup_symbol(const char* name) {
-  for (int i = 0; i < sizeof(table) / sizeof(table[0]); ++i) {
+  for (size_t i = 0; i < sizeof(table) / sizeof(table[0]); ++i) {
     if (strcmp(name, table[i].name) == 0) {
       return table[i].symbol;
     }
@@ -57,7 +57,7 @@ extern "C" MORPHIZEN_DLL_SPEC void* morphizen_main(int argc, char* argv[]) {
   auto cmd = std::string(argv[0]);
   if (cmd == "help") {
     std::cout << "Available commands:" << std::endl;
-    for (int i = 0; i < sizeof(table) / sizeof(table[0]); ++i) {
+    for (size_t i = 0; i < sizeof(table) / sizeof(table[0]); ++i) {
       std::cout << table[i].name << std::endl;
     }
     return nullptr;
